upd + docs + refactor + fix
