+++ .idea/sqldialects.xml
@@ -1,7 +1,6 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <project version="4">
   <component name="SqlDialectMappings">
-    <file url="file://$PROJECT_DIR$/src/main.py" dialect="GenericSQL" />
     <file url="file://$PROJECT_DIR$/src/requesto/requesto.py" dialect="GenericSQL" />
   </component>
 </project>
