--- docs/en/welcome.md
@@ -1,42 +1,0 @@
-**requesto-py** is a lightweight, stable and smart Python library performing ORM functions.
-<br>
-Documentation is !current for the version: `1.2.0-r`
----
-
-<h3>Features:</h3>
-<br>
-* High speed of the program due to optimized code
-* Ease of use due to a reorientation database for OOP programming
-* Unique system of objects due to algorithm of reading tables of database
-* Good for open-source and small projects
-* Open-source
-
----
-## Getting started
-
-### **First steps**: [**Introduction**](./start.md) | [**Quickstart**](./quickstart.md)
-
-
----
-
-## Getting help
-
-**<p>If you're having trouble with something, these resources might help.</p>**
-* **Try to find solution in ** [**Frequently asked questions**](./faq.md)
-* **In case of specific question** [**check out our detailed documentation**](./manuals.md)
-* **Ask it in our GitHub discussion page** [**github**](https://github.com/SOLIDusr/requesto-py/issues/new)
-* **Report bugs and issues in** [**issue tracker**](https://github.com/SOLIDusr/requesto-py/issues/new)
-
----
-
-## Instructions
-**<p>Pages with explanation of different parts of the library</p>**
-* [**requesto.py**](manuals.md)
-
-
----
-
-## Extra
-
-nothing to do
-
