--- docs/lang.md
@@ -1,11 +1,0 @@
-## Choose your language | Выберите язык
-
-<h3>
-  
-[English](./en/welcome.md) 
-
-<br>
-
-[Русский](./ru/welcome.md)
-
-</h3>
