--- docs/ru/welcome.md
@@ -1,40 +1,0 @@
-**requesto-py** это легкая, стабильная и удобная Python библиотека, исполняющая функции ORM.
-<br>
-Документация !актуальна для версии: `1.1.0-r1`
----
-
-<h3>Особенности:</h3>
-<br>
-* Высокая скорость работы за счет оптимизированного кода
-* Удобство использования за счет переориентации работы с базой данных на ООП
-* Простота использования в Open-source и малых проектах
-* Открытый код
-
----
-## Начало
-
-### **Первые шаги**: [**Знакомство**](./start.md) | [**Быстрый старт**](./quickstart.md)
-
----
-
-## Помощь
-
-**<p>Если вы столкнулись с проблемами во время разработки:</p>**
-* **Попробуйте найти решение проблемы в** [**Часто задаваймые вопросы**](./faq)
-* **В случае чего-то специфичного** [**загляните в подробную документацию**](./manuals.md)
-* **Создайте обсуждение на** [**github**](https://github.com/SOLIDusr/requesto-py/issues/new)
-* **Отправьте отчет о баге на** [**issue tracker**](https://github.com/SOLIDusr/requesto-py/issues/new)
-
----
-
-## Инструкции
-**<p>Страницы с описанием действия различных частей библиотеки</p>**
-* [**Manuals**](manuals.md)
-
-
----
-
-## Extra
-
-nothing to do
-
