--- tests/Test Results - Linux Mint 21.2.xml
@@ -1,29 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?><testrun duration="68" footerText="Generated by PyCharm on 16.01.2024, 11:13" name="">
-    <count name="total" value="1"/>
-    <count name="passed" value="1"/>
-    <config configId="tests" name="">
-        <module name="requesto-py"/>
-        <option name="ENV_FILES" value=""/>
-        <option name="INTERPRETER_OPTIONS" value=""/>
-        <option name="PARENT_ENVS" value="true"/>
-        <option name="SDK_HOME" value="/home/naromori/requesto-py/venv/bin/python"/>
-        <option name="WORKING_DIRECTORY" value="/home/naromori/requesto-py/tests/unit"/>
-        <option name="IS_MODULE_SDK" value="true"/>
-        <option name="ADD_CONTENT_ROOTS" value="true"/>
-        <option name="ADD_SOURCE_ROOTS" value="true"/>
-        <EXTENSION ID="PythonCoverageRunConfigurationExtension" runner="coverage.py"/>
-        <option name="_new_keywords" value="&quot;&quot;"/>
-        <option name="_new_parameters" value="&quot;&quot;"/>
-        <option name="_new_additionalArguments" value="&quot;-s&quot;"/>
-        <option name="_new_target" value="&quot;columnNames.test_1_pg_insert&quot;"/>
-        <option name="_new_targetType" value="&quot;PYTHON&quot;"/>
-        <method v="2"/>
-    </config>
-    <suite duration="68" locationUrl="python&lt;/home/naromori/requesto-py/tests/unit&gt;://columnNames" name="columnNames" status="passed">
-        <test duration="68" locationUrl="python&lt;/home/naromori/requesto-py/tests/unit&gt;://columnNames.test_1_pg_insert" name="test_1_pg_insert" metainfo="test_1_pg_insert" status="passed">
-            <output type="stdout">Input Database password
-dolta@83.220.170.127(coredb)$ PASSED
-</output>
-        </test>
-    </suite>
-</testrun>
